Move experimental classes to own modules
