+++ src/oemof/solph/buses/experimental/__init__.py
@@ -7,4 +7,8 @@ Requirements for documentation and unit tests are relaxed,
 so code referred to within this module might not have production quality.
 """
 
-from ._electrical_line import ElectricalLine  # noqa: F401
+from ._electrical_bus import ElectricalBus
+
+__all__ = [
+    "ElectricalBus",
+]
