+++ src/oemof/solph/components/experimental/_generic_caes.py
