+++ src/oemof/solph/components/experimental/_link.py
