+++ src/oemof/solph/components/experimental/_sink_dsm.py
@@ -27,7 +27,7 @@ from pyomo.environ import NonNegativeReals
 from pyomo.environ import Set
 from pyomo.environ import Var
 
-from oemof.solph.components import Sink
+from oemof.solph.components._sink import Sink
 from oemof.solph._options import Investment
 from oemof.solph._plumbing import sequence
