+++ src/oemof/solph/flows/__init__.py
@@ -11,11 +11,10 @@ Note that only mature code is imported,
 experimental code should be included in oemof.experimental.
 """
 
+from . import experimental
 from ._flow import Flow
 
-from . import experimental
-
 __all__ = [
-    "Flow",
     "experimental",
+    "Flow",
 ]
