+++ src/oemof/solph/flows/experimental/_electrical_line.py
@@ -25,9 +25,8 @@ from pyomo.environ import Set
 from pyomo.environ import Var
 
 from oemof.solph._plumbing import sequence as solph_sequence
-from oemof.solph.buses._electrical_bus import ElectricalBus
-
-from ._flow import Flow
+from oemof.solph.flows._flow import Flow
+from oemof.solph.buses.experimental._electrical_bus import ElectricalBus
 
 
 class ElectricalLine(Flow):
